fix(deployment-ui): Replace table with a list
