+++ packages/deployment-server-ui/package.json
@@ -24,6 +24,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.2.1",
+    "react-transition-group": "^4.4.2",
     "react-use": "^17.3.2"
   },
   "devDependencies": {
