--- packages/deployment-server-ui/src/components/Dashboard/RecentDeployments.css
@@ -1,4 +1,0 @@
-.ant-table-thead > tr > th {
-  background: #fff;
-  border-bottom: solid 1px;
-}
