+++ packages/deployment-server-ui/src/pages/Dashboard.tsx
@@ -1,5 +1,5 @@
 import { Divider } from "antd";
-import RecentDeployments from "../components/Dashboard/RecentDeployments";
+import RecentDeploymentList from "../components/Dashboard/RecentDeploymentList";
 import defineFunctionalComponent from "../util/defineFunctionalComponent";
 
 export default defineFunctionalComponent(function Dasboard() {
@@ -7,7 +7,7 @@ export default defineFunctionalComponent(function Dasboard() {
     <main>
       <h2> Dashboard </h2>
       <Divider />
-      <RecentDeployments />
+      <RecentDeploymentList />
     </main>
   );
 });
